Refactoring
Moving all information in the wordpack.ts in to the contextpack.ts.
