+++ client/src/app/contextpack.service.spec.ts
@@ -2,8 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { HttpClient } from '@angular/common/http';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { ContextpackService } from './contextpack.service';
-import { ContextPack } from './contextpack/contextpack';
-import { WordPack, Words } from './wordpack/wordpack';
+import { ContextPack, WordPack, Words } from './contextpack/contextpack';
 
 // Test Context Packs courtesy of the purple tigers @ https://github.com/UMM-CSci-3601-S21/it-1-purple-tigers
 describe('ContextpackService', () => {
